+++ src/utils/websiteUrl.js
@@ -1,0 +1,3 @@
+const BASE_URL = 'https://kafka.js.org'
+
+module.exports = (path, hash) => `${BASE_URL}/${path}${hash ? '#' + hash : ''}`